Bootstrap the baseline model

Run a subset of the baseline model a number of times after resampling e.g. Hauls in each Stratum, EDUSs in each Stratum.

Bootstrap(
  outputData,
  projectPath,
  BootstrapMethodTable = data.table::data.table(),
  NumberOfBootstraps = 1L,
  OutputProcesses = character(),
  UseOutputData = FALSE,
  NumberOfCores = 1L,
  BaselineSeedTable = data.table::data.table()
)

Arguments

outputData
The output of the function from an earlier run.

projectPath
The path to the project to containing the baseline to bootstrap.

BootstrapMethodTable
A table of the columns ProcessName, ResampleFunction and Seed, where each row defines the resample function to apply to the output of the given process (and the seed to use in the resampling). Run RstoxFramework::getResampleFunctions() to get a list of the implemented resample functions. Note that if a process is selected inn BootstrapMethodTable that is not used in the model up to the OutputProcesses, the bootstrapping of that process will not be effective on the end result (e.g. select the correct process that returns BioticAssignment data type).

NumberOfBootstraps
Integer: The number of bootstrap replicates.

OutputProcesses
A vector of the processes to save from each bootstrap replicate.

UseOutputData
Logical: If TRUE use the existing output file from the function.

NumberOfCores
The number of cores to use for parallel processing. A copy of the project is created in tempdir() for each core, also when using only one core. Note that this will require disc space equivalent to the NumberOfCores time the size of the project folder (excluding the output/analysis/Bootstrap folder, which will be deleted before copies are made).

BaselineSeedTable
A table of ProcessName and Seed, givng the seed to use for the Baseline processes that requires a Seed parameter.

Value

A BootstrapData object, which is a list of the RstoxData DataTypes and RstoxBase DataTypes.

Details

A copy of the project is made for each core given by NumberOfCores. In the case that NumberOfCores == 1, this is still done for safety.