StoX data types of the RstoxFramework package

StoX data types are the data types used to transfer data and information between processes in a StoX estimation model. The data types are divided into two types, the ModelData and ProcessData.

Details

The StoX output data can be written to file using the "Write output to file" process parameter in the StoX GUI which corresponds to the logical parameter fileOutput in the project description, which is set to TRUE by default. The output files are written differently based on the class of the output data from the process, with the following file types/file extensions:

Output data class	StoX data types	Output file type	Output file extension

NetCDF4	Bootstrap	NetCDF4	nc

Spatial (multi)polygon. R class: sf	StratumPolygon	GeoJSON	geojson

Table (R class: data.table)	ReadBiotic, StoxBiotic, NASC, etc	Tab separated, NA as "NA"	txt

Plot. R class: ggplot	PlotReportBootstrap, etc	Default: PNG (user choice)	Default: png (user choice)

Character matrix	WriteICESAcoustic, WriteICESBiotic, etc	Comma separated, NA as ""	csv

Vector	WriteICESDatras, WriteICESDatsusc, etc	Comma separated	csv

Output data cannot be composed of multiple output data classes.