Plot ReportBootstrapData

Plot ReportBootstrapData

PlotReportBootstrap(
  ReportBootstrapData,
  PlotType = c("ErrorBarPlot"),
  PlottingVariable = character(),
  PlottingVariableLower = character(),
  PlottingVariableUpper = character(),
  AddCVToPlot = FALSE,
  CVVariable = character(),
  GroupingVariables = character(),
  SubPlots = character(),
  UseDefaultTextSettings = TRUE,
  Title = character(),
  AxisTitleSize = numeric(),
  AxisTickSize = numeric(),
  LegendTitleSize = numeric(),
  LegendTextSize = numeric(),
  UseDefaultFileSettings = TRUE,
  Format = character(),
  Width = numeric(),
  Height = numeric(),
  DotsPerInch = numeric()
)

Arguments

ReportBootstrapData

The ReportBootstrapData data.

PlotType

The type of plot to produce. Currently only "ErrorBarPlot" is implemented.

PlottingVariable

The variable to plot the points for. Normally this is the variable giving the mean.

PlottingVariableLower

The variable to use for the lower end of the error bars.

PlottingVariableUpper

The variable to use for the upper end of the error bars.

AddCVToPlot

Logical: If TRUE add a dotted line with points for the coefficient of variation (standard deviation divided by mean) CV. Setting this to TRUE requires to specify the CVVariable.

CVVariable

The name of the variable holding the CV. Must be given if AddCVToPlot is TRUE.

GroupingVariables

A vector of length 1 or 2 giving the names of the variables to plot along the x axis, where the second is coded with colors in the plot.

SubPlots

A vector of the names of the plots to produce. To get a list of possible plot names use getSubPlotNames_PlotReportBootstrap(ReportBootstrapData, GroupingVariables).

UseDefaultTextSettings

Logical: If TRUE (default) use the default text settings of the plotting function. Setting this to FALSE will show all text options of the plotting function in the StoX GUI.

Title

The title of the plot, defaulted to no title.

AxisTitleSize

The size of the axis titles, defaulted to 20.

AxisTickSize

The size of the axis tick marks, defaulted to 20.

LegendTitleSize

The size of the legend title, defaulted to 20.

LegendTextSize

The size of the legend text, defaulted to 20.

UseDefaultFileSettings

Logical: If TRUE (default) use the default file settings of the plotting function. Setting this to FALSE will show all file options of the plotting function in the StoX GUI.

Format

The format of the output file from plotting functions, defaulted to "png". Other options are defined by ggsave, including "tiff", "jpeg" and "pdf".

Width

The width of the plot in centimeters, defaulted to 17, as per the ICES Journal og Marine Science author recommendations.

Height

The height of the plot in centimeters, defaulted to 17.

DotsPerInch

The number of dots per inch (DPI), defaulted to 1200.

Value

A PlotReportBootstrapData object.