Create, open, close, save, saveAs, copy and delete a StoX project.

Create a StoX project using createProject; open an existing (un-opened) project using openProject, which involves creating files holding the memory of the project; close a project using closeProject, which removes the memory files; save the project using saveProject, which saves the memory files to the project description file; make a copy using copyProject or saveAsProject, where the former closes the given project unsaved and opens the copy, or delete a project using deleteProject.

createProject(
  projectPath,
  template = "EmptyTemplate",
  ow = FALSE,
  showWarnings = FALSE,
  open = TRUE,
  Application = R.version.string
)

openProject(
  projectPath,
  showWarnings = FALSE,
  force = FALSE,
  reset = FALSE,
  type = getRstoxFrameworkDefinitions("projectDescriptionFileFormats"),
  saveIfAlreadyOpen = FALSE,
  verbose = FALSE
)

closeProject(
  projectPath,
  save = NULL,
  Application = R.version.string,
  msg = TRUE
)

saveProject(
  projectPath,
  type = getRstoxFrameworkDefinitions("projectDescriptionFileFormats"),
  force = FALSE,
  Application = R.version.string,
  msg = TRUE
)

saveAsProject(
  projectPath,
  newProjectPath,
  ow = FALSE,
  Application = R.version.string
)

copyProject(projectPath, newProjectPath, ow = FALSE)

deleteProject(projectPath)

Arguments

projectPath
The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.

template
A string naming the template to use when generating the project. See getAvaiableTemplates for a list of available templates.

ow
Logical: If TRUE overwrite the project.

showWarnings
Logical: If TRUE display warninigs when creting the project folders.

open
Logical: If TRUE open the project after creating it.

Application
A single string naming the application used when saving the project. Defaulted to R.version.string.

force
Logical: If TRUE reopen (close and then open) the project if already open.

reset
Logical: If TRUE reset each model to the start of the model.

type
The type of file to save the project to.

saveIfAlreadyOpen
Logical: If TRUE save the project before closing if already open and force is TRUE.

verbose
Logical: If TRUE, print information to the console, e.g. about backward compatibility.

save
Logical: If TRUE save the project before closing. Default (NULL) is to ask the user whether to save the project before closing.

msg
Logical: If FALSE no messages are printed to console (except possibly for extremely important ones).

newProjectPath
The path to the copied StoX project.