Function for comparing existing output files with the memory read using runProject()

Function for comparing existing output files with the memory read using runProject()

compareProjectToStoredOutputFilesAll(
  projectPaths,
  projectPaths_original = projectPaths,
  emptyStringAsNA = FALSE,
  intersect.names = TRUE,
  ignore.variable = NULL,
  skipNAFraction = FALSE,
  skipNAAt = FALSE,
  NAReplacement = NULL,
  classOf = c("first", "second"),
  try = TRUE,
  data.out = FALSE,
  ...
)

Arguments

projectPaths

The projects to be run and tested against the existing output files of the project gievn by projectPath_original.

projectPaths_original

The projects holding the existing output files, defaulted to projectPath.

emptyStringAsNA

Logical: If TRUE, read empty strings as NA from the stored original tables, as RstoxFramework has started writing NAs as NAs and not as empty strings.

intersect.names

Logical: If TRUE, compare only same named columns.

ignore.variable

Either a vector of names of variables/columns to ignore in the comparison, or a list of such variables per process (named by the process names).

skipNAFraction

Logical: If TRUE, skip rows with more than 50 percent NAs. Can also be set to a numeric value between 0 and 1.

skipNAAt

A vector of strings naming the columns in which NA values identifies rows to skip. If more than one variable is given and more than one of these are present in a table, all rows where at least one of the variables are NA are skipped. Note that this may reduce the number of rows and may results in diffs for that reason. Using this option is best used in combination with mergeWhenDifferentNumberOfRows.

NAReplacement

List of replacement values for different classes of NA, applied after any merging as to incorporate NAs generated during merging.

classOf

Character string specifying whether to compare after converting to the class of the first or second table. Set this to "first" (default) to convert class to the original data.

try

Logical: If FALSE do not run the process in a tryCatch. Set this to FALSE when debugging, as the tryCatch masks the errors in the traceback.

data.out

Logical, if TRUE output the original and new data along with the tests. data.out = NULL implies data.out = FALSE if no difference was found and data.out = TRUE otherwise.

...

Arguments passed to readBootstrapData, e.g. selection, which must be set to NA to read the entire file.