Function to convert data.table to fixed width:

Function to convert data.table to fixed width:

fixedWidthDataTable(
  x,
  columnSeparator = " ",
  lineSeparator = NULL,
  na = "-",
  list.pretty = FALSE
)

Arguments

x
The table to modify.

columnSeparator
The string to separate columns by, defaulted to a single space.

lineSeparator
The string to separate lines by, defaulted to a NULL, which keeps the output as a vector of strings.

na
The string to replace NAs by, defaulted to "-".

list.pretty
Logical: If TRUE wrap the output in a list in the case that pretty is TRUE.