General parameters of RstoxFramework.

All functions referring to a project, a model, a process or an output table use the same parameters, listed here.

Arguments

projectPath
The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.

modelName, modelNames
The name of the model(s) (possible values are "baseline", "analysis" and "report").

processID
The ID of the process.

processName, processes
The name of the process(es).

tableName
The name of the output table to get from the process.

startProcess, endProcess
The process index, name or ID at which to start and stop the model run.

afterProcessID, beforeProcessID
The ID of the process after or before which to get the procecss table (place a process or in the case of afterProcessID).

warn
Logical: If TRUE show warnings that are not highly important.

verbose
Logical: If TRUE extra messages are printed to console.

msg
Logical: If FALSE no messages are printed to console (except possibly for extremely important ones).

archive
Logical: Should the project memory state be archived using archiveProject? This stores one step in the history.

template
A string naming the template to use when generating the project. See getAvaiableTemplates for a list of available templates.

ow
Logical: If TRUE overwrite the project.

showWarnings
Logical: If TRUE display warninigs when creting the project folders.

Application
A single string naming the application used when saving the project. Defaulted to R.version.string.

argumentFilePaths
A nested list of paths to argument files of a model, as returned from getArgumentFilePaths. This is used to speed up some functions.

only.valid
Logical: If TRUE subset function arguments (inputs and parameters) to only those to be shown as a consequence of argument hierarchy (e.g., one argument being irrelevant for a specific setting of another).

returnProcessTable
Logical: If TRUE return the process table (much used in functions used by the GUI).