Get output of a StoX process.

Gets the output of a process that has been run.

getProcessOutput(
  projectPath,
  modelName,
  processID,
  tableName = NULL,
  subFolder = NULL,
  flatten = FALSE,
  pretty = FALSE,
  pageindex = integer(0),
  linesPerPage = 1000L,
  columnSeparator = " ",
  lineSeparator = NULL,
  na = "-",
  enable.auto_unbox = TRUE,
  drop = FALSE,
  drop.datatype = TRUE
)

Arguments

projectPath
The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.

modelName
The name of the model (one of "baseline", "analysis" and "report").

processID
The ID of the process.

tableName
The name of the table to extract from the process.

subFolder
If the process returns subfolders (ReadBiotic and ReadAcoustic, where the subfolders represent files), specify the name of the folder with this parameter.

flatten
Logical: Should the output tables that contain cells of length > 1 be expanded to that the other columns are repeated, resulting in a regular table.

pretty
Logical: If TRUE pad with space in each cell to the maximum number of characters of the column including header.

pageindex
A vevctor of the pages to return with linesPerPage number of lines (rows). Default is to not split into pages.

linesPerPage
The number of lines per page if pageindex is given.

columnSeparator
The string to separate columns by, defaulted to a single space.

lineSeparator
The string to separate lines by, defaulted to a NULL, which keeps the output as a vector of strings.

na
The string to replace NAs by, defaulted to "-".

enable.auto_unbox
Logical: If TRUE wrap the output in a list if pretty is TRUE and the output is of length 1. This keeps the array when converting to JSON also for length 1.

drop
Logical: If TRUE drop the list if only one element.

drop.datatype
Logical: If TRUE drop the top level of the output if in a list, which is the level named by the data type.