Get bootstrap data saved in a NetCDF4 file

Get bootstrap data saved in a NetCDF4 file

readBootstrapData(
  filePath,
  nc,
  selection = list(),
  BootstrapID = NA,
  unlistSingleTable = FALSE,
  close = TRUE
)

Arguments

filePath

The path to the file.

nc

A netCDF4 object, overiding the filePath.

selection

Hierarchical list of names of the groups/variables. The last element must be a vector of the variables to return from the table specified by the other elements. E.g., list("ImputeSuperIndividuals", "SuperIndividualsData", c("Stratum", "IndividualAge", "Abundance")) will return a data.table of the three columns "Stratum", "IndividualAge" and "Abundance", added the BootstrapID specified in BootstrapID.

BootstrapID

A sequence of bootstrap IDs, i.e., the indices of the bootstrap replicates. The default returns all bootstrap replicates.

unlistSingleTable

Logical: For use when only single table process outputs are among the requested processes in OutputProcesses of Bootstrap. If FALSE (default) return a list named by the selection with a sub-list named by the datatype holding the output data (e.g. $ImputeSuperIndividuals$SuperIndividualsData). If TRUE return a list named by the selection holding the output data, replicating the output of Bootstrap in StoX <= 3.6.2.

close

Logical: If TRUE (the default) close the netCDF4 file after reading the data.