Read the output files of a project

This function reads all or some of the output files of a project, indicated by model and proccess names.

readModelData(
  projectPath,
  modelName = NULL,
  processName = NULL,
  verifyFiles = FALSE,
  unlist.models = FALSE
)

Arguments

projectPath
The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.

modelName
The name of the model(s) (possible values are "baseline", "analysis" and "report").

processName
The name of the process(es).

verifyFiles
Logical: If TRUE verify that the files are from processes that exist in the project.

unlist.models
Logical: If TRUE unlist the top level so that all processes are in one list.

Value

A list of model output.