Function to read a single process output file, possibly by pages and in flattened and pretty view:

Function to read a single process output file, possibly by pages and in flattened and pretty view:

readProcessOutputFile(
  filePath,
  flatten = FALSE,
  pretty = FALSE,
  pretty.json = TRUE,
  pageindex = integer(0),
  linesPerPage = 1000L,
  columnSeparator = " ",
  lineSeparator = NULL,
  na = "-",
  enable.auto_unbox = FALSE,
  splitGeoJson = TRUE,
  add.line.index = TRUE,
  returnBootstrapData = FALSE,
  selection = list(),
  BootstrapID = NA,
  unlistSingleTable = FALSE
)

Arguments

filePath

The file path of the process output file to read.

flatten

Logical: Should the output tables that contain cells of length > 1 be expanded to that the other columns are repeated, resulting in a regular table.

pretty

Logical: If TRUE pad with space in each cell to the maximum number of characters of the column including header.

pretty.json

Logical: If TRUE prettify the geojson.

pageindex

A vector of the pages to return with linesPerPage number of lines (rows). Default is to not split into pages.

linesPerPage

The number of lines per page if pageindex is given.

columnSeparator

The string to separate columns by, defaulted to a single space.

lineSeparator

The string to separate lines by, defaulted to a NULL, which keeps the output as a vector of strings.

na

The string to replace NAs by, defaulted to "-".

enable.auto_unbox

Logical: If TRUE wrap the output in a list if pretty is TRUE and the output is of length 1. This keeps the array when converting to JSON also for length 1.

splitGeoJson

Logical: If TRUE split the geojson into a vector of separate lines.

add.line.index

Logical: If TRUE (the default) print row indices as in data.table.

returnBootstrapData

Logical: If TRUE read the content of bootstrap NetCDF4 file.

selection

Hierarchical list of names of the groups/variables. The last element must be a vector of the variables to return from the table specified by the other elements. E.g., list("ImputeSuperIndividuals", "SuperIndividualsData", c("Stratum", "IndividualAge", "Abundance")) will return a data.table of the three columns "Stratum", "IndividualAge" and "Abundance", added the BootstrapID specified in BootstrapID.

BootstrapID

A sequence of bootstrap IDs, i.e., the indices of the bootstrap replicates. The default returns all bootstrap replicates.

unlistSingleTable

Logical: For use when only single table process outputs are among the requested processes in OutputProcesses of Bootstrap. If FALSE (default) return a list named by the selection with a sub-list named by the datatype holding the output data (e.g. $ImputeSuperIndividuals$SuperIndividualsData). If TRUE return a list named by the selection holding the output data, replicating the output of Bootstrap in StoX <= 3.6.2.