Run processes of a model.

Run processes of a model.

runProcesses(
  projectPath,
  modelName,
  startProcess = 1,
  endProcess = Inf,
  msg = TRUE,
  msg.GUI = FALSE,
  save = TRUE,
  force.save = FALSE,
  saveProcessData = TRUE,
  Application = R.version.string,
  force.restart = FALSE,
  returnProcessOutput = FALSE,
  fileOutput = NULL,
  setUseProcessDataToTRUE = TRUE,
  purge.processData = FALSE,
  replaceDataList = list(),
  replaceArgsList = list(),
  prependProcessList = list(),
  try = TRUE,
  prugeStopFile = FALSE,
  ...
)

Arguments

projectPath

The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.

modelName

The name of the model (possible values are "baseline", "analysis" and "report").

startProcess

The process index, name or ID at which to start the model run.

endProcess

The process index, name or ID at which to stop the model run.

msg

Logical: If FALSE no messages are printed to console (except possibly for extremely important ones).

msg.GUI

Logical: If TRUE, exlcude certain messages (used by the StoX GUI).

save

Logical: If TRUE save the project before closing. Default (NULL) is to ask the user whether to save the project before closing.

force.save

If no changes are made to the project, force save anyway. Overrides the save option.

saveProcessData

Logical: If TRUE save the output as ProcessData, if the proecss is a ProcessData process.

Application

A single string naming the application used when saving the project. Defaulted to R.version.string.

force.restart

Logical: If TRUE, start the processes even if the status file indicating that the model is being run exists. This is useuful when something crached in a preivous run, in which case the model is still appearing as running.

returnProcessOutput

Logical: If TRUE return the process output immediately after it is available. Used to get filter options.

fileOutput

Logical: If TRUE save the output as a text file (or other format specified by the class or attributes of the output). If NULL (defafult) use the corresponding parameter of the process.

setUseProcessDataToTRUE

Logical: If TRUE set the UseProcessData function parameter to TRUE in the process memory after execution, if the process is a ProcessData process.

purge.processData

Logical: If TRUE replace process data entirely.

replaceDataList

A list named by the processes to replace output data for. See runProcess.

replaceArgsList

A list of replaceArgs (see runProcess) holding parameters to replace in the function call, named by the processes to modify.

prependProcessList

A list of values used in prependProcess, named by the processes to prepend a process to.

try

Logical: If FALSE do not run the process in a tryCatch. Set this to FALSE when debugging, as the tryCatch masks the errors in the traceback.

prugeStopFile

Logical: Should the file that signals that the model should be stopped be deleted if present before running (used by the GUI).

...

replaceArgsList can also be given directly.