Run all models of a StoX project

This function runs and returns output from all models of a StoX project.

runProject(
  projectPath,
  modelNames = getRstoxFrameworkDefinitions("stoxModelNames"),
  processes = NULL,
  startProcess = 1,
  endProcess = Inf,
  drop.datatype = TRUE,
  unlistDepth2 = FALSE,
  run = TRUE,
  save = TRUE,
  force.restart = FALSE,
  replaceDataList = list(),
  replaceArgs = list(),
  fileOutput = NULL,
  setUseProcessDataToTRUE = TRUE,
  purge.processData = FALSE,
  try = TRUE,
  close = FALSE,
  unlist.models = TRUE,
  ...
)

Arguments

projectPath
The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.

modelNames
The name of the model(s) (possible values are "baseline", "analysis" and "report").

processes
The name of the process(es).

startProcess
The process index, name or ID at which to start and stop the model run.

endProcess
The process index, name or ID at which to start and stop the model run.

drop.datatype
Logical: If TRUE, drop the top level of the output list if it has length 1 and that single element is named by the datatype name.

unlistDepth2
Logical: Related to drop.datatype, but setting this to TRUE unlists output data that are nested in 2 levels, such as output from ReadBiotic, which outputs a set of tables for each input file. Using unlistDepth2 = TRUE puts all these tables in one list, and uses the concatenation of the file names and the table name separated by underscore. This is how it is displayed in the StoX GUI when selecting "View output".

run
Logical: If TRUE run the model.

save
Logical: If TRUE save the project before closing. Default (NULL) is to ask the user whether to save the project before closing.

force.restart
Logical: If TRUE, start the processes even if the status file indicating that the model is being run exists. This is useuful when something crached in a preivous run, in which case the model is still appearing as running.

replaceDataList
A list named by the processes to replace output data for. See runProcess.

replaceArgs
A list of function parameters and inputs to override.

fileOutput
Logical: If TRUE save the output as a text file (or other format specified by the class or attributes of the output). If NULL (defafult) use the corresponding parameter of the process.

setUseProcessDataToTRUE
Logical: If TRUE set the UseProcessData function parameter to TRUE in the process memory after execution, if the process is a ProcessData process.

purge.processData
Logical: If TRUE replace process data entirely.

try
Logical: If FALSE do not run the process in a tryCatch. Set this to FALSE when debugging, as the tryCatch masks the errors in the traceback.

close
Logical: If TRUE close the project after running and getting the output.

unlist.models
Logical: If TRUE unlist the top level so that all processes are in one list.

...
replaceArgs can also be given directly.

Value

A list of model output.