Write/append a table/list of tables to NetCDF4

Write/append a table/list of tables to NetCDF4

write_list_as_tables_NetCDFF4(
  list,
  filePath,
  nc,
  index,
  dims,
  nchars,
  append = FALSE,
  ow = FALSE,
  missval = -9,
  compression = NA,
  verbose = FALSE,
  validOutputDataClasses = getRstoxFrameworkDefinitions("validOutputDataClasses")
)

Arguments

list

A list of tables to write to NetCDF4.

filePath

The path to the file.

nc

A netCDF4 object, overiding the filePath.

index

The index of the current data to write, with number of rows indicated in the dims argument.

dims

A list of lists of dimensions of the tables to write.

nchars

A list of lists of lists holding the maximum number of characters for each variable of each table of each process.

append

Logical: If TRUE, append to the existing file.

ow

Logical: If TRUE, overwrite the existing file.

missval

The value to use for representing missing values (NA) for numeric variables.

compression

If set to an integer between 1 (least compression) and 9 (most compression), this enables compression for the variable as it is written to the file. Turning compression on forces the created file to be in netcdf version 4 format, which will not be compatible with older software that only reads netcdf version 3 files.

verbose

Print debugging information.

validOutputDataClasses

A vector of valid output data classes, indicating when to stop the unlisting.