Create, open, close, save, saveAs, copy and delete a StoX project.

Create a StoX project using createProject; open an existing (un-opened) project using openProject, which involves creating files holding the memory of the project; close a project using closeProject, which removes the memory files; save the project using saveProject, which saves the memory files to the project description file; make a copy using copyProject or saveAsProject, where the former closes the given project unsaved and opens the copy, or delete a project using deleteProject.

createProject(
  projectPath,
  template = "EmptyTemplate",
  ow = FALSE,
  showWarnings = FALSE,
  open = TRUE,
  Application = R.version.string
)

openProject(
  projectPath,
  showWarnings = FALSE,
  force = FALSE,
  reset = FALSE,
  saveIfAlreadyOpen = FALSE,
  verbose = FALSE
)

openProjectAsTemplate(
  projectPath,
  newProjectPath,
  close = FALSE,
  save = NULL,
  ow = FALSE
)

closeProject(
  projectPath,
  save = NULL,
  force.save = FALSE,
  Application = R.version.string,
  msg = TRUE
)

saveProject(
  projectPath,
  force = FALSE,
  Application = R.version.string,
  msg = TRUE
)

saveAsProject(
  projectPath,
  newProjectPath,
  ow = FALSE,
  Application = R.version.string
)

copyProject(
  projectPath,
  newProjectPath,
  ow = FALSE,
  empty.output = FALSE,
  empty.input = FALSE,
  empty.memory = FALSE,
  empty.processData = FALSE,
  processDataToBeEmptied = NULL,
  close = FALSE,
  save = NULL,
  msg = TRUE
)

deleteProject(projectPath)

getProject(projectPath)

Arguments

projectPath

The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.

template

A string naming the template to use when generating the project. See getAvaiableTemplates for a list of available templates.

ow

Logical: If TRUE overwrite the project.

showWarnings

Logical: If TRUE display warninigs when creting the project folders.

open

Logical: If TRUE open the project after creating it.

Application

A single string naming the application used when saving the project. Defaulted to R.version.string.

force

Logical: If TRUE reopen (close and then open) the project if already open.

reset

Logical: If TRUE reset each model to the start of the model.

saveIfAlreadyOpen

Logical: If TRUE save the project before closing if already open and force is TRUE.

verbose

Logical: If TRUE, print information to the console, e.g. about backward compatibility.

newProjectPath

The path to the copied StoX project.

close

Logical: (In copyProject) If TRUE, close the project before copying.

save

Logical: If TRUE save the project before closing. Default (NULL) is to ask the user whether to save the project before closing.

force.save

If no changes are made to the project, force save anyway. Overrides the save option.

msg

Logical: If FALSE no messages are printed to console (except possibly for extremely important ones).

empty.output

Logical: If TRUE, do not include the output files when copying. This can also be a vector of names of the output folders to empty.

empty.input

Logical: If TRUE, do not include the input files when copying. This can also be a vector of names of the input data folders to empty.

empty.memory

Logical: If TRUE, do not include the memory data files when copying. This can also be a vector of names of the memory data folders to empty.

empty.processData

Logical: (In copyProject) If TRUE, empty the process data listed in processDataToBeEmptied.

processDataToBeEmptied

An optional character vector of names of the process data to empty when copying a StoX project. The default (NULL) implies all possible process data (all functions returned from getRstoxFrameworkDefinitions("processDataFunctions")).