Function for comparing existing output files with the memory read using runProject()

Function for comparing existing output files with the memory read using runProject()

compareProjectToStoredOutputFiles(
  projectPath,
  projectPath_original = projectPath,
  emptyStringAsNA = FALSE,
  intersect.names = TRUE,
  ignore = NULL,
  skipNAFraction = FALSE,
  skipNAAt = FALSE,
  setNATo0 = FALSE,
  classOf = c("first", "second"),
  try = TRUE
)

Arguments

projectPath

The project to be run and tested against the existing output files of the project gievn by projectPath_original.

projectPath_original

The project holding the existing output files, defaulted to projectPath.

emptyStringAsNA

Logical: If TRUE, read empty strings as NA from the stored original tables, as RstoxFramework has started writing NAs as NAs and not as empty strings.

intersect.names

Logical: If TRUE, compare only same named columns.

ignore

A vector of names of columns to ignore in all.equal().

skipNAFraction

Logical: If TRUE, skip rows with more than 50 percent NAs. Can be set to a value between 0 and 1.

skipNAAt

A vector of strings naming the columns in which NA values identifies rows to skip.

setNATo0

Logical: If TRUE, set all NAs to 0.

classOf

Character string specifying whether to compare after converting to the class of the first or second table. Set this to "first" (default) to convert class to the original data.

try

Logical: If FALSE do not run the process in a tryCatch. Set this to FALSE when debugging, as the tryCatch masks the errors in the traceback.