Compare processes in a StoX 2.7 project and a copy in StoX >= 3.0.0:

Compare processes in a StoX 2.7 project and a copy in StoX >= 3.0.0:

compareSweptAreaBaseline(
  projectPathOld,
  projectPathNew,
  comparisonModel,
  output_Old = NULL,
  output_New = NULL,
  data.out = FALSE
)

Arguments

projectPathOld

The path to the StoX 2.7 project.

projectPathNew

The path to the copy project in StoX >= 3.0.0.

comparisonModel

A list with specifications of what and how to compare the outputs from the old and new project (see Details).

output_Old

The StoX 2.7 table.

output_New

The StoX 2.7 table.

data.out

The StoX 2.7 table.

The comparisonModel is a list with the following required elements:

dataType

The StoX >= 3.0.0 datatypes to compare. Used only for reference.

dataVariable_Old

The specific variable to compare in the old datatypes.

dataVariable_New

The specific variable to compare in the new datatypes.

processName_Old,processName_New

The process named in the old and new project, respectively. Must correspond to the dataType.

compareAction

A string vector specifyfing whether to merge ("merge") or cbind ("cbind") when comparing. Cbind is only for SuperIndividuals in StoX 2.7, which does not have unique keys in.

keys_Old,keys_New"

A list of the keys of the datatypes in the old and new project, respectively. Specify only for datatypes to be merged (use NA for those that use compareAction = "cbind").

subsetByNAOn_Old,subsetByNAOn_New"

A list of variables to subset the old and new data by, respectively. When any of these are NA in a row, the row is skipped. Specify only for datatypes to be cbinded (use NA for those that use compareAction = "merge").

orderBy_Old,orderBy_New

SA list of variables to order the old and new data by, respectively (after subsetting using subsetByNAOn_Old and subsetByNAOn_New). Specify only for datatypes to be cbinded (use NA for those that use compareAction = "merge").

tolerance

A numeric vector setting the tolerance for the comparison, below which the outputs are regarded as equal.