General parameters of RstoxFramework.

All functions referring to a project, a model, a process or an output table use the same parameters, listed here.

Arguments

projectPath

The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.

projectPaths

A vector of paths to the StoX projects.

projectDescriptionFile

The path to the file holding the projectDescription (project.json).

modelName

The name of the model (possible values are "baseline", "analysis" and "report").

modelNames

The name of the models (possible values are "baseline", "analysis" and "report").

processID

The ID of the process.

processName

The name of the process.

processNames

A vector of names of the processes.

processes

The name of the processes to return data from.

functionName

The name of the function used by the process.

tableName

The name of the output table to get from the process.

startProcess

The process index, name or ID at which to start the model run.

endProcess

The process index, name or ID at which to stop the model run.

beforeProcessID

The ID of the process before which to get the process table.

afterProcessID

The ID of the process after which to get the process table or to place a process.

warn

Logical: If TRUE show warnings that are not highly important.

verbose

Logical: If TRUE extra messages are printed to console.

msg

Logical: If FALSE no messages are printed to console (except possibly for extremely important ones).

msg.GUI

Logical: If TRUE, exlcude certain messages (used by the StoX GUI).

archive

Logical: Should the project memory state be archived using archiveProject? This stores one step in the history.

template

A string naming the template to use when generating the project. See getAvaiableTemplates for a list of available templates.

ow

Logical: If TRUE overwrite the project.

showWarnings

Logical: If TRUE display warninigs when creting the project folders.

Application

A single string naming the application used when saving the project. Defaulted to R.version.string.

argumentFilePaths

A nested list of paths to argument files of a model, as returned from getArgumentFilePaths. This is used to speed up some functions.

only.valid

Logical: If TRUE subset function arguments (inputs and parameters) to only those to be shown as a consequence of argument hierarchy (e.g., one argument being irrelevant for a specific setting of another).

returnProcessTable

Logical: If TRUE return the process table of the specified model (much used in functions used by the GUI). If NULL returns the process table of all models.

applyBackwardCompatibility

Logical: If TRUE apply backward compatibility actions when running readProjectDescription.

Seed

The seed, given as a single integer.