Get output of a StoX process.

Gets the output of a process that has been run.

getProcessOutput(
  projectPath,
  modelName,
  processID,
  tableName = NULL,
  subFolder = NULL,
  flatten = FALSE,
  pretty = FALSE,
  pretty.json = FALSE,
  pageindex = integer(0),
  linesPerPage = 1000L,
  columnSeparator = " ",
  lineSeparator = NULL,
  na = "-",
  enable.auto_unbox = TRUE,
  drop = FALSE,
  drop.datatype = TRUE,
  splitGeoJson = TRUE,
  warn = TRUE,
  add.line.index = TRUE,
  returnBootstrapData = FALSE,
  selection = list(),
  BootstrapID = NA,
  unlistSingleTable = FALSE,
  unlistSingleBootstrapData = TRUE
)

getProcessTableOutput(
  projectPath,
  modelName,
  processID,
  tableName = NULL,
  flatten = FALSE,
  pretty = FALSE,
  pageindex = integer(0),
  linesPerPage = 1000L,
  columnSeparator = " ",
  na = "-",
  drop = FALSE,
  add.line.index = TRUE
)

getProcessGeoJsonOutput(
  projectPath,
  modelName,
  processID,
  geoJsonName = NULL,
  pretty = FALSE,
  splitGeoJson = TRUE
)

Arguments

projectPath

The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.

modelName

The name of the model (possible values are "baseline", "analysis" and "report").

processID

The ID of the process.

tableName

The name of the table to extract from the process.

subFolder

If the process returns subfolders (ReadBiotic and ReadAcoustic, where the subfolders represent files), specify the name of the folder with this parameter.

flatten

Logical: Should the output tables that contain cells of length > 1 be expanded to that the other columns are repeated, resulting in a regular table.

pretty

Logical: If TRUE pad with space in each cell to the maximum number of characters of the column including header.

pretty.json

Logical: If TRUE prettify the geojson.

pageindex

A vector of the pages to return with linesPerPage number of lines (rows). Default is to not split into pages.

linesPerPage

The number of lines per page if pageindex is given.

columnSeparator

The string to separate columns by, defaulted to a single space.

lineSeparator

The string to separate lines by, defaulted to a NULL, which keeps the output as a vector of strings.

na

The string to replace NAs by, defaulted to "-".

enable.auto_unbox

Logical: If TRUE wrap the output in a list if pretty is TRUE and the output is of length 1. This keeps the array when converting to JSON also for length 1.

drop

Logical: If TRUE drop the list if only one element.

drop.datatype

Logical: If TRUE drop the top level of the output if in a list, which is the level named by the data type.

splitGeoJson

Logical: If TRUE split the geojson into a vector of separate lines.

warn

Logical: If TRUE warn if the process has not bee run.

add.line.index

Logical: If TRUE (the default) print row indices as in data.table.

returnBootstrapData

Logical: If TRUE read the content of bootstrap NetCDF4 file.

selection

Hierarchical list of names of the groups/variables. The last element must be a vector of the variables to return from the table specified by the other elements. E.g., list("ImputeSuperIndividuals", "SuperIndividualsData", c("Stratum", "IndividualAge", "Abundance")) will return a data.table of the three columns "Stratum", "IndividualAge" and "Abundance", added the BootstrapID specified in BootstrapID.

BootstrapID

A sequence of bootstrap IDs, i.e., the indices of the bootstrap replicates. The default returns all bootstrap replicates.

unlistSingleTable

Logical: For use when only single table process outputs are among the requested processes in OutputProcesses of Bootstrap. If FALSE (default) return a list named by the selection with a sub-list named by the datatype holding the output data (e.g. $ImputeSuperIndividuals$SuperIndividualsData). If TRUE return a list named by the selection holding the output data, replicating the output of Bootstrap in StoX <= 3.6.2.

unlistSingleBootstrapData

Logical: If TRUE unlist the the top level of the process output if this is a list with a single element of class "BootstrapData" (and named "BootstrapData"). This is the default for historic reasons to match the output of StoX < 4.0.0.

geoJsonName

The name of the GeoJSON object to extract from the process.