Functions to get the process table of a model.

Functions to get the process table of a model.

getProcessTable(
  projectPath,
  modelName = NULL,
  startProcess = 1,
  endProcess = Inf,
  afterProcessID = NULL,
  beforeProcessID = NULL,
  argumentFilePaths = NULL,
  only.valid = TRUE,
  return.processIndex = FALSE
)

scanForModelError(
  projectPath,
  modelName = NULL,
  startProcess = 1,
  endProcess = Inf,
  afterProcessID = NULL,
  beforeProcessID = NULL,
  argumentFilePaths = NULL,
  only.valid = TRUE,
  return.processIndex = FALSE
)

getProcessesSansProcessData(
  projectPath,
  modelName = NULL,
  startProcess = 1,
  endProcess = Inf,
  afterProcessID = NULL,
  beforeProcessID = NULL,
  argumentFilePaths = NULL,
  only.valid = FALSE,
  return.processIndex = FALSE
)

getProcessAndFunctionNames(
  projectPath,
  modelName = NULL,
  startProcess = 1,
  endProcess = Inf,
  afterProcessID = NULL,
  beforeProcessID = NULL,
  argumentFilePaths = NULL,
  return.processIndex = FALSE
)

Arguments

projectPath

The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.

modelName

The name of the model (possible values are "baseline", "analysis" and "report").

startProcess

The process index, name or ID at which to start the model run.

endProcess

The process index, name or ID at which to stop the model run.

afterProcessID

The ID of the process after which to get the process table or to place a process.

beforeProcessID

The ID of the process before which to get the process table.

argumentFilePaths

A nested list of paths to argument files of a model, as returned from getArgumentFilePaths. This is used to speed up some functions.

only.valid

Logical: If TRUE subset function arguments (inputs and parameters) to only those to be shown as a consequence of argument hierarchy (e.g., one argument being irrelevant for a specific setting of another).

return.processIndex

Logical: If TRUE include the process indices as a sequence starting from 1 in each model.